+++ client/CHANGELOG.md
@@ -21,6 +21,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Changed Add Row button styling in Advanced Search ([#642](https://github.com/project-lux/lux-frontend/issues/642)).
+
 ### Removed
 
 ### Fixed